+++ app/components/layer-chooser/layer-chooser-directives.js
@@ -1,71 +1,71 @@
 //layer-directive.js
 
-angular.module('lizard-nxt')
-  .directive("layerChooser", ['NxtMap', 'NxtData', 'dataLayers', 'State', 'DataService',
-    function (NxtMap, NxtData, dataLayers, State, DataService) {
+angular.module('data-menu')
+  .directive("layerChooser", ['State', 'LeafletService',
+    function (State, LeafletService) {
 
   var link = function (scope, element, attrs) {
     // Scope gets the mapState layerGroup, here we create a new layerGroup which
     // goes into its own NxtMap to always be turned on
-    var layerGroup = angular.copy(dataLayers[scope.layergroup.slug]);
+    var layerGroup = scope.layergroup;
 
-    var needsPreviewMap = false;
-    // Remove unnecessary datalayers before going further
-    // but first copying the layer array, otherwise the
-    // iterator gets confused.
-    var layerCopy = angular.copy(layerGroup.layers);
-    angular.forEach(layerGroup.layers, function (layer) {
-      if (layer.format === 'WMS' ||
-          layer.format === 'TMS') {
-        needsPreviewMap = true;
-        layerCopy = [layer];
+    var leafletLayers = [];
+
+    angular.forEach(layerGroup.mapLayers, function (layer) {
+      if (layer.format === 'TMS') {
+        leafletLayers.push(LeafletService.tileLayer(
+          layer.url + '/{slug}/{z}/{x}/{y}.{ext}', {
+            slug: layer.slug,
+            minZoom: layer.min_zoom || 0,
+            maxZoom: 19,
+            detectRetina: true,
+            zIndex: layer.zIndex,
+            ext: 'png'
+          }));
       }
-
-      if (layer.format === 'Vector') {
+      else if (layer.format === 'WMS') {
+        var _options = {
+          layers: layer.slug,
+          format: 'image/png',
+          version: '1.1.1',
+          minZoom: layer.min_zoom || 0,
+          maxZoom: 19,
+          opacity: layer.opacity,
+          zIndex: layer.zIndex
+        };
+        _options = angular.extend(_options, layer.options);
+        leafletLayers.push(LeafletService.tileLayer.wms(layer.url, _options));
+      }
+      else if (layer.format === 'Vector') {
         element.find('.layer-img')[0].style.backgroundColor = layer.color;
       }
     });
 
-    layerGroup.layers = layerCopy;
+    if (leafletLayers.length > 0) {
+      var map = L.map(element.find('.layer-img')[0], {
+          center: [52.39240447569775, 5.101776123046875],
+          zoom: 6,
+          dragging: false,
+          touchZoom: false,
+          doubleClickZoom: false,
+          tap: false,
+          scrollWheelZoom: false,
+          animate: true,
+          zoomControl: false,
+          attributionControl: false,
+        }
+      );
 
-    if (needsPreviewMap) {
-      var previewMap = new NxtMap();
-      previewMap.createMap(element.find('.layer-img')[0], {
-        center: [52.39240447569775, 5.101776123046875],
-        zoom: 6,
-        dragging: false,
-        touchZoom: false,
-        doubleClickZoom: false,
-        tap: false,
-        scrollWheelZoom: false,
-        animate: true,
-        zoomControl: false,
-        attributionControl: false,
-        forChooser: true
-      });
-
-      var data = new NxtData([layerGroup]);
-
-      data.syncTime(State.temporal, previewMap._map);
-
-      // To speed-up initial load of the app layerchooser layers are toggled
-      // after 3000 ms.
-      setTimeout(function () {
-        data.toggleLayerGroup(
-          data.layerGroups[Object.keys(data.layerGroups)[0]], previewMap._map
-          );
-      }, 3000);
+      leafletLayers.forEach(function (l) { l.addTo(map); });
 
       scope.$watch(State.toString('spatial.bounds'), function (n, v) {
         if (n === v) { return; }
         var zoom = State.spatial.zoom;
         var centroid = State.spatial.bounds.getCenter();
-        previewMap.setView(centroid, zoom - 2);
-        scope.bounds = State.spatial.bounds;
+        map.setView(centroid, zoom - 2);
       });
     }
 
-    scope.data = DataService;
   };
 
   return {