+++ app/lib/raster-service.js
@@ -2,8 +2,10 @@
  * Service to handle raster requests.
  */
 angular.module('lizard-nxt')
-  .service("RasterService", ["Restangular", "UtilService", "CabinetService", "$q",
-
+  .service("RasterService", ["Restangular",
+                             "UtilService",
+                             "CabinetService",
+                             "$q",
   function (Restangular, UtilService, CabinetService, $q) {
 
   var intensityData,
@@ -12,7 +14,7 @@ angular.module('lizard-nxt')
   var getData = function (layer, options) {
 
     // TODO: get this from somewhere
-    var GRAPH_WIDTH = window.innerwidth;
+    var GRAPH_WIDTH = UtilService.getCurrentWidth();
 
     var srs = 'EPSG:4326',
         agg = options.agg || '',
@@ -29,15 +31,16 @@ angular.module('lizard-nxt')
     aggWindow = options.aggWindow || UtilService.getAggWindow(options.start,
       options.end, GRAPH_WIDTH);
 
+    var canceler;
     // getData can have own deferrer to prevent conflicts
     if (options.deferrer) {
-      var deferSlug = options.deferrer.origin,
-          canceler = options.deferrer.deferred;
+      var deferSlug = options.deferrer.origin;
+      canceler = options.deferrer.deferred;
       if (cancelers[options.deferrer.origin]) {
         cancelers[options.deferrer.origin].resolve();
       }
       cancelers[options.deferrer.origin] = canceler;
-    } 
+    }
     // if it doesn't have a deferrer in the options
     // use the layer slug..
       else {
@@ -45,7 +48,7 @@ angular.module('lizard-nxt')
         cancelers[layer.slug].resolve();
       }
 
-      var canceler = cancelers[layer.slug] = $q.defer();
+      canceler = cancelers[layer.slug] = $q.defer();
     }
 
     return CabinetService.raster(canceler).get({
@@ -55,6 +58,7 @@ angular.module('lizard-nxt')
       start: startString,
       stop: endString,
       agg: agg,
+      styles: options.styles,
       window: aggWindow
     });
   };
@@ -117,13 +121,8 @@ angular.module('lizard-nxt')
   return {
     getMinTimeBetweenFrames: getMinTimeBetweenFrames,
     buildURLforWMS: buildURLforWMS,
-    // rasterInfo: rasterInfo,
-    //getRasterData: getRasterData,
     getData: getData,
-    // getTemporalRaster: getTemporalRaster,
-    // getImgOverlays: getImgOverlays,
     handleElevationCurve: handleElevationCurve,
-    //getRasterDataForExtentData: getRasterDataForExtentData,
   };
 
 }]);