+++ test/karma.conf.dev.js
@@ -37,6 +37,9 @@ module.exports = function(config) {
       // application
       'test/mocks/beforeModuleMocks.js',
       'app/components/state/state.js', // Load these first to prevent dep clash
+      'app/components/data-menu/data-menu.js', // Load these first to prevent dep clash
+      'app/components/map/map.js', // Load these first to prevent dep clash
+      'app/components/state/state.js', // Load these first to prevent dep clash
       'app/templates.js',
       'app/components/omnibox/omnibox.js',
       'app/lizard-nxt.js',